[FE] add : youMayLike 컴포넌트 제작
